Make fromObject used by Variant and FromObject classes handle AttributeError
This could occur when passing None to fromUnicode(), something we
always do as a final fallback, even if the type isn't a str/bytes.
This is more likely now as we attempt more conversions.

We catch the error instead of doing an 'if not None' check to allow
for fields that *want* to handle None values in fromUnicode.

Also provide more useful reprs for VariantValidationError:

```
VariantValidationError()
    Value: None
    Field: <nti.schema.field.Variant at 0x10b2666d8 name='' interface=None context=None fields=[<nti.schema.field.HTTPURL object at 0x10b266a58>]>
      zope.schema._bootstrapinterfaces.WrongType: (None, <class 'str'>, '')`
```
`